Highlights
Dynamically redirect Flows automatically or through a button on the screen using a Record Id or a URL.
Launch a Flow from a button on a Lightning Page and refresh the record when the Flow completes.
Unlock and Lock records through Flow and Process Builder. Show Users if a record is locked or unlocked with the Locked Record Notifier Lightning Component.